+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 
 - `impl Clone for OneDrive`.
+- Configurable `Tenant` for `Auth`, this allows login into applications that
+  accept only personal accounts.
 
 ### Changed